feat: add auto loading and create custom hook with main logic
meta: issue-1_logic
