+++ package.json
@@ -12,6 +12,7 @@
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.18",
     "graphql": "^16.8.1",
+    "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
@@ -42,5 +43,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/lodash": "^4.14.202"
   }
 }
