+++ src/hooks/useArticleList.tsx
@@ -1,0 +1,55 @@
+import { useEffect, useState } from 'react';
+import { useQuery } from '@apollo/client';
+import { ARTICLES_QUERY } from '../utils';
+import { Article } from '../types';
+import { debounce } from 'lodash';
+
+export const useArticleList = () => {
+  const { loading, error, data, fetchMore } = useQuery(ARTICLES_QUERY, {
+    variables: { skip: 0 },
+    fetchPolicy: 'cache-and-network',
+  });
+
+  const [articles, setArticles] = useState<Article[]>([]);
+  const [loadingMore, setLoadingMore] = useState<boolean>(false);
+
+  useEffect(() => {
+    if (!loading && data && data.contents) {
+      setArticles(data.contents);
+    }
+  }, [loading, data]);
+
+  const loadMore = async () => {
+    if (loadingMore) return; // Предотвращаем дублирование запросов
+    try {
+      setLoadingMore(true);
+      const result = await fetchMore({
+        variables: {
+          skip: articles.length,
+        },
+      });
+      const newArticles = result.data.contents;
+      setArticles(prevArticles => [...prevArticles, ...newArticles]);
+    } catch (err) {
+      console.error('Error loading more articles:', err);
+    } finally {
+      setLoadingMore(false);
+    }
+  };
+
+  const handleScroll = debounce(() => {
+    const { scrollTop, clientHeight, scrollHeight } = document.documentElement;
+    if (scrollTop + clientHeight >= scrollHeight - 500) {
+      loadMore();
+    }
+  }, 100);
+
+  useEffect(() => {
+    window.addEventListener('scroll', handleScroll);
+    return () => {
+      window.removeEventListener('scroll', handleScroll);
+    };
+  }, [handleScroll]);
+
+  return { loading, error, articles, loadingMore };
+};
